Add additional OOP classes
