+++ src/main/java/Command.java
@@ -1,0 +1,5 @@
+public abstract class Command {
+    public abstract void execute(TaskList taskList, Ui ui, Storage storage);
+
+    public abstract boolean isExit();
+}
