+++ src/main/java/ExitCommand.java
@@ -1,0 +1,8 @@
+public class ExitCommand extends Command{
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        ui.showExit();
+    }
+    public boolean isExit() {
+        return true;
+    }
+}
