+++ src/main/java/ListTasksCommand.java
@@ -1,0 +1,8 @@
+public class ListTasksCommand extends Command{
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        ui.showTaskListContents(taskList.getTaskStore());
+    }
+    public boolean isExit() {
+        return false;
+    }
+}
